feat: add DocWriter, SimpleWriter, UnitExt, move elem
